Improved layout of inlets and outlets.
The control and audio inlets are now separate, and the control outlet is
placed on the left. Previously the leftmost inlet was used both for control
and audio input, and the control outlet was on the right, which had several
disadvantages.

The new layout is symmetrical between in- and outputs and thus looks tidier,
it simplifies the code in faust_tilde_io.c, and has the control outlet always
in the same place, so that it doesn't have to be recreated (losing current
connections) when the number of outputs changes. Also, it avoids the confusing
signal inlet on the left if there are no audio inputs. Incidentally, this is
also the layout used by pd-faust, which makes it easier to port patches from
pd-faust to pd-faustgen.

Also fixed a memory leak in faust_io_manager_free, which previously didn't
free the control outlet when a faustgen~ object was destroyed.