Add support for automatic Pd GUI generation.
This works pretty much like in pd-faust, i.e., you first create a one-off
subpatch and then specify the name of the subpatch as the instance name of the
faustgen~ object. The contents of the subpatch will then be populated with GUI
controls for the Faust dsp, setting a suitable GOP area. The subpatch will
also be updated each time the object gets recompiled, and a special 'gui'
message is provided to manually trigger the GUI update if needed.