Add basic tuning support.
Octave-based tunings can now be set with the 'tuning' message. Note that these
will only have an effect on polyphonic dsps, where they determine the tuning
offsets in cent for each note in the octave relative to the default equal
temperament. The following variations of the tuning message are provided:

- tuning: Outputs the current tuning on the control outlet (this will be
  'tuning default' for the default 12-tet tuning).

- tuning default: Resets the tuning to the default 12-tet.

- tuning offset ...: Sets the tuning offsets in cent for each tone in the
  octave (thus 12 numbers need to be supplied as arguments).

- tuning scala-filename [base]: Constructs a tuning from the given Scala
  file (http://www.huygens-fokker.org/scala/). The .scl extension is supplied
  automatically. Only tunings with 12 scale tones are recognized. If base is
  given, it must denote the reference tone (0-11) which gets a tuning offset
  of 0. The default base is 0, denoting the first scale tone (C); specify a
  base of 9 to fix the tuning of A4 at 440 Hz.

An abundance of different scale files is readily available from the Scala
website, and you can easily create your own if needed. See:
http://www.huygens-fokker.org/scala/scl_format.html