Reworked the MIDI channel configuration.
We're using a proper bitmask for up to 64 different MIDI channels now, so that
multiple individual channels can be specified either in the creation arguments
or in the midichan message. As before, 0 denotes omni. Negative channel
numbers can now be used to denote specific channels that are to be
blocked (e.g., midichan -10 will block the GM drumkit channel).

A midichan message without any arguments now outputs the current setup on the
control outlet; as before, you can use midichan 0 to set omni mode.

We also removed the symbolic "omni" and "gm" creation arguments, to prevent
potential clashes with instance names.