Add a panic control.
Polyphonic dsps now accept a bang message which turns off all sounding notes.
This is also available on the Pd GUI as a third button (the gray bang control
in the top row), and gives you a quick way to turn off hanging notes.
(Incidentally, that button also indicates that the corresponding dsp is a
polyphonic one.)

Note that faustgen~ doesn't interpret CC 123 (the all-notes-off controller) by
default, but you can get that functionality now by hooking up a [ctlin 123]
object to a bang message that is sent to a synth receiver, or just the global
faustgen~ receiver to be on the safe side (non-polyphonic dsps will just
ignore the bang message).