Basic OSC support.
This adds a handler for OSC messages of the form `/msg data ...`, where `/msg`
is the OSC address a.k.a. message selector, and `data ...` is a sequence of
one or more floating point values. Using the "alias" mechanism described in
the OSC Support section of the Faust manual, OSC addresses can be assigned to
controls by meta data of the form `[osc:/msg]`, optionally adding an OSC range
`a b` for the control for automatic mapping purposes (the default range, if a
and b are not specified, is 0 1). This works with both active Faust
controls (allowing input of controls by OSC messages) and passive ones (output
of OSC messages to the control outlet and/or a receiver symbol).

On input, multi-argument messages of the form `/msg arg0 arg1 ...` are handled
by sending the different values to the OSC addresses /msg/0, /msg/1, etc., as
described in the Faust manual.

In a fashion analogous to the midiout creation argument and message, the
oscout creation argument and message can be used to control OSC output.
oscout=num enables or disables OSC output to the control outlet, depending on
whether num is nonzero or zero (the latter is the default). At runtime, the
oscout num message does the same. Also, oscout=sym (or oscout sym as a
message) sends OSC output to the given receiver symbol sym. Finally, the
oscout message without any arguments disables output to the current receiver
symbol.

This implementation merely provides a basic framework for getting OSC messages
into and out of a faustgen~ object. To interface to external OSC devices,
you'll have to use Pd's OSC input/output facilities, such as those provided by
the OSCx and mrpeach externals.