Add a /dump message.
This message provides a simple OSC inspection capability. It is by no means as
comprehensive as the elaborate interface described in the OSC Support section
of the Faust manual. But it provides the most essential functionality to
inspect the OSC interface of a faustgen~ object, including its OSC values and
ranges, while also keeping things simple.

When received by a faustgen~ object, /dump without any arguments outputs the
current values of the OSC-enabled UI elements to the control outlet and/or OSC
receiver symbol, as a sequence of /msg data OSC messages. This form of the
message is useful, e.g., to provide parameter feedback to connected OSC
devices.

When invoked with a single symbol argument, say /dump /dump.reply, /dump
outputs a sequence of OSC messages /dump.reply /msg data a b, specifying the
message selector, current value, and range of each OSC-enabled UI element.
The sequence of output messages is terminated by the OSC message
/dump.reply done to indicate the end of the transmission. This form of the
message is useful for discovering the OSC interface of a faustgen~ object, so
it may be used by OSC applications, e.g., to generate a default UI for the
object.