Add a Pd loader extension for Faust dsp files.
This means that to create a dsp object you can now just type the name of the
dsp (without the .dsp extension) followed by the usual creation arguments,
without explicitly using the faustgen~ object (which still works, too). The
dsp file will be searched in the usual places. Note that to enable the loader
extension, faustgen~ needs to be loaded at startup, so you want to put it into
your startup libraries to make this work.

The dsp name can be followed by a ~, which is customary notation for dsp
objects in Pd land, so we support that. This is purely cosmetic, though, you
can always omit the trailing ~, and in any case it will be ignored when
searching for the file. (You can still load a dsp file with a trailing ~ in
its name, but you'll have to go through the faustgen~ object if you want to do
that.)