Update the help patch and the example patches.
We're now using the loader extension throughout, which makes the patches look
tidier and will also help when we rename the faustgen~ object. Which is in
order because we want to be able to use the original CICM version and the JGU
version of the external at the same time.