Reworked the 'set parameter list' interface so that it works again.
This previously required literal blank characters in symbols, which stopped
working when we introduced symbol mangling to turn arbitrary Faust labels into
legal Pd symbols. We changed the interface now so that symbols of the form
name%i are being used, which don't cause any problems in Pd.

Also fixed some minor glitches in the set.parameters.lists example in the help
patch.